services/game-catalog/requirements.txt
#TODO: Completare
flask
pymongo
flask-bcrypt
flask-jwt-extended
kafka-python